+++ src/Services/CRM/Automation/Batch.php
@@ -1,0 +1,164 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\CRM\Automation;
+
+use Bitrix24\SDK\Core\Commands\Command;
+use Bitrix24\SDK\Core\Commands\CommandCollection;
+use Bitrix24\SDK\Core\Contracts\CoreInterface;
+use Bitrix24\SDK\Core\Exceptions\BaseException;
+use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
+use Bitrix24\SDK\Core\Response\DTO\Pagination;
+use Bitrix24\SDK\Core\Response\DTO\ResponseData;
+use Bitrix24\SDK\Core\Response\DTO\Result;
+use Bitrix24\SDK\Core\Response\DTO\Time;
+use Bitrix24\SDK\Core\Response\Response;
+use Generator;
+use Psr\Log\LoggerInterface;
+
+/**
+ * Class Batch
+ *
+ * @package Bitrix24\SDK\Services\CRM\Automation
+ */
+class Batch extends \Bitrix24\SDK\Core\Batch
+{
+    /**
+     * Delete entity items with batch call
+     *
+     *
+     * @return Generator<int, ResponseData>|ResponseData[]
+     * @throws \Bitrix24\SDK\Core\Exceptions\BaseException
+     */
+    public function deleteEntityItems(
+        string $apiMethod,
+        array $entityItemId,
+        ?array $additionalParameters = null
+    ): Generator {
+        $this->logger->debug(
+            'deleteEntityItems.start',
+            [
+                'apiMethod' => $apiMethod,
+                'entityItems' => $entityItemId,
+                'additionalParameters' => $additionalParameters,
+            ]
+        );
+
+        try {
+            $this->clearCommands();
+            foreach ($entityItemId as $cnt => $code) {
+                if (!is_string($code)) {
+                    throw new InvalidArgumentException(
+                        sprintf(
+                            'invalid type «%s» of trigger code «%s» at position %s, code must be string type',
+                            gettype($code),
+                            $code,
+                            $cnt
+                        )
+                    );
+                }
+
+                $parameters = ['CODE' => $code];
+                $this->registerCommand($apiMethod, $parameters);
+            }
+
+            foreach ($this->getTraversable(true) as $cnt => $deletedItemResult) {
+                yield $cnt => $deletedItemResult;
+            }
+        } catch (InvalidArgumentException $exception) {
+            $errorMessage = sprintf('batch delete entity items: %s', $exception->getMessage());
+            $this->logger->error(
+                $errorMessage,
+                [
+                    'trace' => $exception->getTrace(),
+                ]
+            );
+            throw $exception;
+        } catch (\Throwable $exception) {
+            $errorMessage = sprintf('batch delete entity items: %s', $exception->getMessage());
+            $this->logger->error(
+                $errorMessage,
+                [
+                    'trace' => $exception->getTrace(),
+                ]
+            );
+
+            throw new BaseException($errorMessage, $exception->getCode(), $exception);
+        }
+
+        $this->logger->debug('deleteEntityItems.finish');
+    }
+
+    /**
+     * Execute triggers with batch call
+     *
+     *
+     * @return Generator<int, ResponseData>|ResponseData[]
+     * @throws \Bitrix24\SDK\Core\Exceptions\BaseException
+     */
+    public function executeItems(
+        string $apiMethod,
+        array $triggers
+    ): Generator {
+        $this->logger->debug(
+            'executeItems.start',
+            [
+                'apiMethod' => $apiMethod,
+                'triggers' => $triggers,
+            ]
+        );
+
+        try {
+            $this->clearCommands();
+            foreach ($triggers as $cnt => $fields) {
+                if (!is_array($fields)) {
+                    throw new InvalidArgumentException(
+                        sprintf(
+                            'invalid type «%s» of trigger params «%s» at position %s, params must be array type',
+                            gettype($fields),
+                            $fields,
+                            $cnt
+                        )
+                    );
+                }
+
+                $this->registerCommand($apiMethod, $fields);
+            }
+
+            foreach ($this->getTraversable(true) as $cnt => $executeResult) {
+                yield $cnt => $executeResult;
+            }
+        } catch (InvalidArgumentException $exception) {
+            $errorMessage = sprintf('batch execute triggers: %s', $exception->getMessage());
+            $this->logger->error(
+                $errorMessage,
+                [
+                    'trace' => $exception->getTrace(),
+                ]
+            );
+            throw $exception;
+        } catch (\Throwable $exception) {
+            $errorMessage = sprintf('batch execute triggers: %s', $exception->getMessage());
+            $this->logger->error(
+                $errorMessage,
+                [
+                    'trace' => $exception->getTrace(),
+                ]
+            );
+
+            throw new BaseException($errorMessage, $exception->getCode(), $exception);
+        }
+
+        $this->logger->debug('executeItems.finish');
+    }
+}